pyproject.toml
[tool.poetry]
name = "paidcontentservice"
version = "0.1.0"
description = ""
authors = ["hoolywant <sir.saltickow@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
stripe = "^7.8.2"
coverage = "^7.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
